Create the method save book

+++ src/main/java/com/marcus/libraryapi/domain/dto/response/BookResponse.java
@@ -15,7 +15,7 @@ public class BookResponse {
     private String isbn;
     private String title;
     private String author;
-    private Long numberPages;
+    private int numberPages;
 
     public static BookResponse of(Book book) {
         return disassemble.apply(book);
